@@ -17,12 +17,12 @@
     "axios": "^0.27.2",
     "bignumber.js": "^9.1.0",
     "formik": "^2.2.9",
-    "framer-motion": "^6.5.1",
+    "framer-motion": "^10.11.2",
     "http-proxy-middleware": "^2.0.6",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-device-detect": "^2.2.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-helmet-async": "^1.3.0",
     "react-popper": "^2.3.0",
     "react-redux": "^8.0.1",
